npm fixes  #1575 npm commands in repl or ui often fail to exit  - sometimes npm process was dying before we could get output. Use WaitHandle instead (this affected both repl and ui)  - npm logic consolidated to single location  - added tests
 #1382 NPM UI Errors out with out-of-memory
 - parse incrementally instead of reading all packages at once.

 Misc. fixes
 - make sure db query doesn't get executed on the UI thread by converting to list after filtering
 - Count instead of LongCount for retrieving ResultsCount

+++ Product/SharedProject/ProcessOutput.cs
@@ -678,7 +678,7 @@ namespace Microsoft.VisualStudioTools.Project {
         /// Immediately stops the process.
         /// </summary>
         public void Kill() {
-            if (_process != null) {
+            if (_process != null && !_process.HasExited) {
                 _process.Kill();
                 // Should have already been called, in which case this is a no-op
                 OnExited(this, EventArgs.Empty);
